Refactor to support more generation types
