+++ generator/codegen/common.go
@@ -13,14 +13,55 @@ import (
 	"golang.org/x/tools/imports"
 
 	"github.com/ARM-software/golang-utils/utils/commonerrors"
+	configUtils "github.com/ARM-software/golang-utils/utils/config"
 	"github.com/ARM-software/golang-utils/utils/filesystem"
 	"github.com/ARM-software/golang-utils/utils/parallelisation"
+	validation "github.com/go-ozzo/ozzo-validation"
 )
 
 var (
 	//go:embed templates/*
 	templates embed.FS
 )
+
+type Data struct {
+	Params          any
+	SpecPath        string
+	TemplatePath    string
+	DestinationPath string
+}
+
+var ValidGenerationTargets = map[string]func(*Data) error{
+	"collections": AddCollectionsToParams,
+}
+
+type ExtensionsConfig struct {
+	Input        string `mapstructure:"input"`
+	Output       string `mapstructure:"output"`
+	Template     string `mapstructure:"template"`
+	GenerateType string `mapstructure:"generate_type"`
+}
+
+func DefaultExtensionsConfig() *ExtensionsConfig {
+	return &ExtensionsConfig{
+		Input:    "",
+		Output:   "",
+		Template: "templates/entities.go.tmpl", // FIXME: delete when CI usage is updated
+	}
+}
+
+func (cfg *ExtensionsConfig) Validate() error {
+	err := configUtils.ValidateEmbedded(cfg)
+	if err != nil {
+		return err
+	}
+
+	return validation.ValidateStruct(cfg,
+		validation.Field(&cfg.Input, validation.Required),
+		validation.Field(&cfg.Output, validation.Required),
+		validation.Field(&cfg.Template, validation.Required),
+	)
+}
 
 func GenerateDataStruct(cfg ExtensionsConfig) (d *Data, err error) {
 	return &Data{
@@ -101,9 +142,29 @@ func generateSourceCode(ctx context.Context, data *Data, template *template.Temp
 	return
 }
 
-func LoadAPIDefinition(specPath string) (swagger *openapi3.T, err error) {
+func loadAPIDefinition(specPath string) (swagger *openapi3.T, err error) {
 	loader := openapi3.NewLoader()
 	loader.IsExternalRefsAllowed = true
 	swagger, err = loader.LoadFromFile(specPath)
 	return
+}
+
+func AddValuesToParams(d *Data, getValues func(*openapi3.T) (interface{}, error), generationType string) (err error) {
+	if d == nil {
+		err = fmt.Errorf("%w input variable", commonerrors.ErrUndefined)
+		return
+	}
+	if d.TemplatePath == "" {
+		d.TemplatePath = fmt.Sprintf("templates/%s.tmpl", generationType)
+	}
+	swagger, err := loadAPIDefinition(d.SpecPath)
+	if err != nil {
+		return
+	}
+	params, err := getValues(swagger)
+	if err != nil {
+		return
+	}
+	d.Params = params
+	return nil
 }
