+++ generator/codegen/templates/entities.go.tmpl
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2020-2024 Arm Limited or its affiliates and Contributors. All rights reserved.
+ * Copyright (C) 2020-2025 Arm Limited or its affiliates and Contributors. All rights reserved.
  * SPDX-License-Identifier: Apache-2.0
  */
